+++ js/util.js
@@ -8,4 +8,4 @@ const getRandomPositiveInteger = (a, b) => {
 const getRandomArrayElement = (array) =>
   array[getRandomPositiveInteger(0, array.length - 1)];
 
-export {getRandomPositiveInteger};
+export {getRandomArrayElement};
